fix(backend): StoreSchedule id 타입을 Long → Int로 수정 — DB serial 컬럼과 일치
displayworkschedulemaster__c 테이블의 id 컬럼이 serial(integer)이므로
Hibernate schema-validation 오류 해결

+++ backend/src/main/kotlin/com/otoki/internal/common/repository/StoreScheduleRepository.kt
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query
 import org.springframework.data.repository.query.Param
 import java.time.LocalDate
 
-interface StoreScheduleRepository : JpaRepository<StoreSchedule, Long> {
+interface StoreScheduleRepository : JpaRepository<StoreSchedule, Int> {
 
     /**
      * 사원(fullName sfid)의 특정 날짜 거래처 스케줄 조회
